backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "sqlalchemy>=2.0.25",
  "passlib[bcrypt]>=1.7.4",
  "chromadb>=0.5.3",
  "sentence-transformers>=3.0.1",
  "langchain-text-splitters>=0.2.2",
  "pypdf>=4.2.0",
  "python-docx>=1.1.0",
  "docx2txt>=0.8",
  "pandas>=2.2.2",
  "httpx>=0.27.0",
  "pillow>=10.3.0",
  "openai>=1.35.0"]

[tool.uvicorn]
factory = false
app = "app.main:app"
host = "0.0.0.0"
port = 8000
reload = true
